Removed CreatedXXX and replaced it with CreatedModel which returns a Model and a BoundingBox.  Fixed VertexDesc layout code to assume a vertex has 3 positions instead of true, which meant only one :(.

+++ SpaceSim/CubeMapRenderer.cpp
@@ -179,7 +179,7 @@ void CubeMapRenderer::renderCubeMap(Resource* resource, Texture* renderTarget,co
                 // Set vertex buffer stride and offset.
                 const VertexBuffer* vb = renderInstance.getGeometryInstance().getVB();
                 ID3D11Buffer* buffer = vb->getBuffer();
-                stride = vb->getVertexStride();
+                stride = static_cast<unsigned int>(vb->getVertexStride());
                 if (renderInstance.getGeometryInstance().getIB() != nullptr)
                 {
                     deviceContext->IASetInputLayout(renderInstance.getGeometryInstance().getVB()->getInputLayout());
@@ -193,7 +193,7 @@ void CubeMapRenderer::renderCubeMap(Resource* resource, Texture* renderTarget,co
                     deviceContext->IASetInputLayout(renderInstance.getGeometryInstance().getVB()->getInputLayout());
                     deviceContext->IASetVertexBuffers(0, 1, &buffer, &stride, &offset);
                     deviceContext->IASetPrimitiveTopology((D3D11_PRIMITIVE_TOPOLOGY)renderInstance.getPrimitiveType());
-                    deviceContext->Draw(renderInstance.getGeometryInstance().getVB()->getVertexCount(), 0);
+                    deviceContext->Draw((unsigned int)renderInstance.getGeometryInstance().getVB()->getVertexCount(), 0);
                 }
             }
         }
